backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
groq==0.4.0
python-dotenv==1.0.0
pydantic==2.4.2
pytest==7.4.3
httpx==0.25.1